Make sure to use /usr/bin/which or equivalent
There is sometimes (certainly in zsh) a shell builtin for which.  It's
behaviour is unpredictable, and the command /usr/bin/which should be
preferred.

Specifically, builtin which in zsh looks through functions as well as
binaries, and so the function

    vim () {
        if which vim; then
            command vim
        else
            command vi
        fi
    }

always runs command vim.

If we want it to use vi if there is no vim binary in the path, we should
instead use `if command which vim; then` etc.

(We could equivalently use `if /usr/bin/which vim; then` etc., but this
is fragile because it assumes that which is always installed at
/usr/bin/which, rather than at any other place.  That's *probably* a
safe assumption, but it's better not to risk it...)